+++ Libplanet/Store/LiteDBStore.cs
@@ -118,6 +118,39 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
+        public IEnumerable<Address> ListAddresses(string @namespace)
+        {
+            var prefix = $"{StateRefIdPrefix}{@namespace}/";
+            foreach (LiteFileInfo fileInfo in _db.FileStorage.Find(prefix))
+            {
+                string fileId = fileInfo.Id;
+                int slashIndex = fileId.LastIndexOf('/');
+                if (slashIndex < 0)
+                {
+                    continue;
+                }
+
+                string addressHex = fileId.Substring(slashIndex + 1);
+                if (addressHex.Length < Address.Size * 2)
+                {
+                    continue;
+                }
+
+                Address address;
+                try
+                {
+                    address = new Address(addressHex);
+                }
+                catch (ArgumentException)
+                {
+                    continue;
+                }
+
+                yield return address;
+            }
+        }
+
+        /// <inheritdoc/>
         public void StageTransactionIds(ISet<TxId> txids)
         {
             StagedTxIds.InsertBulk(
@@ -270,7 +303,7 @@ namespace Libplanet.Store
                 _db.FileStorage.FindById(BlockStateFileId(blockHash));
             if (file is null)
             {
-                return new AddressStateMap();
+                return null;
             }
 
             using (var stream = new MemoryStream())